Move dataset loader tests into new separate file
